Fix implementation to satisfy RFC7230-37
Problem

We currently have two problems.
Firstly, as described in RFC7231, 1XX, 204 and 304 responses
must not include a message-body. This requirement was not enforced in finagle, i.e., a message-body can be added.
Moreover, Content-Length field is added to these responses with value 0, which is not allowed
as mentioned in RFC7231#section-3.3.2

Solution

As I've described we need to solve two things.
First of all, we can replace a message-body(which we call content in finagle) with Buffers.Empty if the response
status code is either 1XX, 204 or 304 to satisfy RFC7231. As response code seems to be finally determined at the point of
Dispatcher, we can do this at that time.
Secondly, we make sure that we do not add Content-Length field even with value 0 to follow RFC7230#section-3.3.2.
This also can be done at the time of dispathcer#handle is called.

Result

We follow the specification of RFC7231 with regard to a message-header and message-body except that we have not added some header fields required to a response with status code 304 with accordance to RFC7232#section-4.1: Cache-Control, Content-Location, Date, ETag, Expires, and Vary.

[1] https://www.rfc-editor.org/rfc/rfc7231.txt
[2] https://tools.ietf.org/html/rfc7230#section-3.3.2
[3] https://tools.ietf.org/html/rfc7232#section-4.1