finagle-http: Fix body and header in 204 and 304
Problem

We currently have two problems. Firstly, as described in RFC2616[1] and
RFC7231[2], 204 and 304 responses must not contain a message-body. This
requirement was not enforced in finagle, i.e., a message-body can be
added. Moreover, a Content-Length header field can be added to these responses,
which is not allowed as mentioned in RFC7230#section-3.3.2[3].

Solution

First of all, we can replace a message-body(which we call content in finagle)
with Buffers.Empty if the response status code is either 204 or 304 to satisfy
RFC7231. As a response status code seems to be finally determined at some point
in HttpServerDispatcher, we can do this in HttpServerDispatcher. Secondly, we
need to make sure not to add a Content-Length header field in a 204 response
to follow RFC7230#section-3.3.2. With regard to a 304 response, a server May
send a Content-Length header as mentioned in the same section. Thus it'd be
better for us to allow users to set the field if they like.

Result

We follow the specification of RFC with regard to a message-header and
message-body except that we have not added some header fields equired in a
response with status code 304 to satisfy RFC7232#section-4.1[4]; Any of the
following header fields must be generated: Cache-Control, Content-Location,
Date, ETag, Expires, and Vary.

[1] https://tools.ietf.org/html/rfc2616
[2] https://tools.ietf.org/html/rfc7231
[3] https://tools.ietf.org/html/rfc7230#section-3.3.2
[4] https://tools.ietf.org/html/rfc7232#section-4.1